Unify UserRequirements in Types and Mutations
